# services/gateway-go/docs/swagger.yaml
basePath: /
definitions:
  openapi.AgentsPage:
    properties:
      data:
        description: Data is the page contents in (block_number DESC, id DESC) order.
        items:
          $ref: '#/definitions/store.Agent'
        type: array
      next_cursor:
        description: NextCursor is the opaque pagination token. Empty on the final
          page.
        type: string
    type: object
  openapi.AuthErrorBody:
    properties:
      code:
        description: |-
          Code is a stable machine-readable identifier such as
          "invalid_signature", "domain_mismatch", "rate_limited".
        example: invalid_signature
        type: string
      message:
        description: |-
          Message is a short human-readable hint. MUST NOT contain values
          derived from the request body (signature bytes, recovered
          address, internal errors).
        example: signature did not verify
        type: string
    type: object
  openapi.AuthErrorResponse:
    properties:
      error:
        $ref: '#/definitions/openapi.AuthErrorBody'
    type: object
  openapi.AuthNonceResponse:
    properties:
      expires_at:
        description: |-
          ExpiresAt is the unix timestamp at which the nonce will be
          rejected by /verify even if signed correctly.
        example: 1714512000
        type: integer
      nonce:
        description: |-
          Nonce is a fresh single-use entropy token bound to the upcoming
          /verify call. Hex-encoded random bytes.
        type: string
    type: object
  openapi.AuthVerifyRequest:
    properties:
      message:
        description: |-
          Message is the EIP-4361 (SIWE) or analogous SIWS message string,
          verbatim — clients MUST NOT pre-parse and re-serialise it.
        type: string
      signature:
        description: |-
          Signature is hex-encoded for SIWE (0x-prefixed, 65 bytes) and
          base58 for SIWS (64-byte ed25519 signature).
        type: string
    type: object
  openapi.AuthVerifyResponse:
    properties:
      address:
        description: |-
          Address is the lowercase 0x-hex EVM address (SIWE) or base58
          Solana pubkey (SIWS) the token was issued to.
        type: string
      expires_at:
        description: ExpiresAt is the unix timestamp the token expires at.
        example: 1714598400
        type: integer
      token:
        description: 'Token is the bearer JWT to attach as Authorization: Bearer <token>.'
        type: string
    type: object
  openapi.HealthResponse:
    properties:
      status:
        description: Status is always "ok" when the gateway is serving.
        example: ok
        type: string
    type: object
  openapi.JobsPage:
    properties:
      data:
        items:
          $ref: '#/definitions/store.Job'
        type: array
      next_cursor:
        type: string
    type: object
  openapi.TradesPage:
    properties:
      data:
        items:
          $ref: '#/definitions/store.Trade'
        type: array
      next_cursor:
        type: string
    type: object
  store.Agent:
    properties:
      agent_id:
        description: on-chain uint256, decimal
        type: string
      block_number:
        type: integer
      capabilities:
        description: bitmap as decimal
        type: string
      controller:
        description: 0x-hex (ACP)
        type: string
      created_at:
        type: string
      creator:
        description: 0x-hex (launchpad)
        type: string
      id:
        type: integer
      kind:
        $ref: '#/definitions/store.AgentKind'
      log_index:
        type: integer
      metadata_uri:
        description: optional
        type: string
      tx_hash:
        description: 0x-hex
        type: string
      updated_at:
        type: string
    type: object
  store.AgentKind:
    enum:
    - launchpad
    - acp
    type: string
    x-enum-varnames:
    - AgentKindLaunchpad
    - AgentKindACP
  store.Job:
    properties:
      agent_id:
        description: on-chain uint256
        type: string
      agent_pk:
        type: integer
      block_number:
        type: integer
      budget:
        type: string
      clone_address:
        type: string
      created_at:
        type: string
      deadline:
        type: string
      id:
        type: integer
      job_type:
        type: integer
      log_index:
        type: integer
      on_chain_id:
        type: string
      phase:
        $ref: '#/definitions/store.JobPhase'
      principal:
        type: string
      result_uri:
        type: string
      token:
        type: string
      tx_hash:
        type: string
      updated_at:
        type: string
    type: object
  store.JobPhase:
    enum:
    - created
    - funded
    - active
    - completed
    - cancelled
    - disputed
    - released
    - resolved
    type: string
    x-enum-varnames:
    - JobPhaseCreated
    - JobPhaseFunded
    - JobPhaseActive
    - JobPhaseCompleted
    - JobPhaseCancelled
    - JobPhaseDisputed
    - JobPhaseReleased
    - JobPhaseResolved
  store.Stats:
    properties:
      last_block_indexed:
        type: integer
      total_agents:
        type: integer
      total_jobs:
        type: integer
      total_trades:
        type: integer
    type: object
  store.Trade:
    properties:
      agent_in:
        type: string
      agent_out:
        type: string
      agent_token_pk:
        type: integer
      block_number:
        type: integer
      created_at:
        type: string
      curve:
        type: string
      fee:
        type: string
      id:
        type: integer
      log_index:
        type: integer
      quote_in:
        type: string
      quote_out:
        type: string
      side:
        description: '"buy" | "sell"'
        type: string
      trader:
        type: string
      tx_hash:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Titular maintainers
    url: https://github.com/0xDevNinja/titular
  description: |-
    Read-only REST surface plus SIWE / SIWS authentication for the
    Titular protocol. The gateway proxies a Postgres-backed read
    model populated by the indexer track and offers a parallel
    GraphQL surface (see /graphql) and SSE event stream
    (see /events). All amount and id fields that originate as
    uint256 on chain are encoded as decimal strings on the wire.
    Authentication uses bearer tokens minted by the SIWE / SIWS
    verify endpoints; the read endpoints are public by default.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/0xDevNinja/titular
  title: Titular Gateway API
  version: 0.1.0-alpha
paths:
  /api/v1/agents:
    get:
      description: |-
        Returns a paginated, opaque-cursor page of registered agents.
        Filterable by agent kind. Cursor is base64-encoded JSON the
        client MUST treat as opaque and pass back unmodified.
      parameters:
      - description: opaque pagination cursor; omit for the first page
        in: query
        name: cursor
        type: string
      - description: page size; 1..200, default 50
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - description: agent kind filter
        enum:
        - launchpad
        - acp
        in: query
        name: kind
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openapi.AgentsPage'
        "400":
          description: unknown query parameter, or kind/limit out of range
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
      summary: List agents
      tags:
      - agents
  /api/v1/agents/{id}:
    get:
      description: |-
        Returns the registered agent identified by its primary key
        (decimal int64) or kind-qualified slug "kind:agent_id".
      parameters:
      - description: agent primary key or kind:agent_id slug
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Agent'
        "404":
          description: agent not found
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
      summary: Get agent
      tags:
      - agents
  /api/v1/jobs:
    get:
      description: |-
        Returns a paginated, opaque-cursor page of ACP jobs.
        Filterable by phase via the `status` query parameter.
      parameters:
      - description: opaque pagination cursor
        in: query
        name: cursor
        type: string
      - description: page size; 1..200, default 50
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - description: ACP job phase filter
        enum:
        - created
        - funded
        - active
        - completed
        - cancelled
        - disputed
        - released
        - resolved
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openapi.JobsPage'
        "400":
          description: unknown query parameter, or invalid status/limit
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
      summary: List ACP jobs
      tags:
      - jobs
  /api/v1/stats:
    get:
      description: |-
        Returns total agent / job / trade counters and the highest
        block number observed by the indexer track. Cached to a few
        seconds upstream; clients SHOULD NOT poll faster than 5s.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Stats'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
      summary: Aggregate protocol stats
      tags:
      - stats
  /api/v1/trades:
    get:
      description: |-
        Returns a paginated, opaque-cursor page of bonding-curve
        trades (buy + sell). Filterable by agent token address and
        by an inclusive [from, to] window on the indexer-recorded
        created_at timestamp.
      parameters:
      - description: opaque pagination cursor
        in: query
        name: cursor
        type: string
      - description: page size; 1..200, default 50
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - description: 0x-prefixed 20-byte token address
        in: query
        name: agent_token
        type: string
      - description: inclusive RFC 3339 lower bound on created_at
        format: date-time
        in: query
        name: from
        type: string
      - description: inclusive RFC 3339 upper bound on created_at
        format: date-time
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openapi.TradesPage'
        "400":
          description: unknown query parameter, or invalid time/limit
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
      summary: List trades
      tags:
      - trades
  /auth/logout:
    post:
      description: |-
        Idempotently revokes the session associated with the
        Authorization bearer token. Returns 204 whether or not the
        session existed; the response is invariant so callers
        cannot probe for live JTIs.
      responses:
        "204":
          description: session revoked or no-op
        "500":
          description: internal_error
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke session
      tags:
      - auth
  /auth/siwe/nonce:
    post:
      description: |-
        Returns a fresh entropy nonce the client embeds in the
        EIP-4361 SIWE message it asks the user to sign. The nonce
        is single-use across both SIWE and SIWS verify endpoints
        and expires after a short server-controlled TTL.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openapi.AuthNonceResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
      summary: Issue SIWE nonce
      tags:
      - auth
  /auth/siwe/verify:
    post:
      consumes:
      - application/json
      description: |-
        Validates an EIP-4361 SIWE message and its EVM signature
        against the previously issued nonce, the gateway's pinned
        domain, and the configured chain id; on success, returns a
        bearer JWT and persists the session in Redis.
      parameters:
      - description: SIWE message and signature
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/openapi.AuthVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openapi.AuthVerifyResponse'
        "400":
          description: invalid_request, invalid_message, domain_mismatch, chain_mismatch,
            expired_message
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
        "401":
          description: invalid_nonce or invalid_signature
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
        "500":
          description: internal_error
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
      summary: Verify SIWE message and mint session
      tags:
      - auth
  /auth/siws/nonce:
    post:
      description: |-
        Returns a fresh entropy nonce the client embeds in the
        SIWS (Sign-In With Solana) message it asks the user to
        sign. The nonce is single-use across both SIWE and SIWS
        verify endpoints and expires after a short server-controlled
        TTL.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openapi.AuthNonceResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
      summary: Issue SIWS nonce
      tags:
      - auth
  /auth/siws/verify:
    post:
      consumes:
      - application/json
      description: |-
        Validates a SIWS (Sign-In With Solana) message and its
        ed25519 signature against the previously issued nonce,
        the gateway's pinned domain, and the configured Solana
        cluster; on success, returns a bearer JWT and persists
        the session in Redis.
      parameters:
      - description: SIWS message and base58 signature
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/openapi.AuthVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openapi.AuthVerifyResponse'
        "400":
          description: invalid_request, invalid_message, domain_mismatch, cluster_mismatch,
            expired_message, invalid_address
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
        "401":
          description: invalid_nonce or invalid_signature
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
        "500":
          description: internal_error
          schema:
            $ref: '#/definitions/openapi.AuthErrorResponse'
      summary: Verify SIWS message and mint session
      tags:
      - auth
  /healthz:
    get:
      description: |-
        Returns {"status":"ok"} as soon as the HTTP server is up.
        This endpoint does NOT depend on Postgres, Redis, or NATS;
        it is intentionally a chassis-only check so orchestrators
        can distinguish "binary running" from "fully wired".
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openapi.HealthResponse'
      summary: Liveness probe
      tags:
      - health
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Bearer token issued by POST /auth/siwe/verify or /auth/siws/verify.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"